Float weights and class_sums

+++ PySparseCoalescedTsetlinMachineCUDA/cuda/code_evaluate.cu
@@ -1,7 +1,7 @@
 #include <curand_kernel.h>
 extern "C" {
 // Evaluate examples
-__global__ void evaluate(unsigned int *global_ta_state, int *clause_weights, int *class_sum, int *X) {
+__global__ void evaluate(unsigned int *global_ta_state, float *clause_weights, float *class_sum, int *X) {
     int index = blockIdx.x * blockDim.x + threadIdx.x;
     int stride = blockDim.x * gridDim.x;
 
@@ -47,7 +47,7 @@ __global__ void evaluate(unsigned int *global_ta_state, int *clause_weights, int
 
         if (clause_output) {
             for (int class_id = 0; class_id < CLASSES; ++class_id) {
-                int clause_weight = clause_weights[class_id * CLAUSES + clause];
+                float clause_weight = clause_weights[class_id * CLAUSES + clause];
                 atomicAdd(&class_sum[class_id], clause_weight);
             }
         }
@@ -56,7 +56,7 @@ __global__ void evaluate(unsigned int *global_ta_state, int *clause_weights, int
 
 // Evaluate examples
 __global__ void evaluate_packed(unsigned int *included_literals, unsigned int *included_literals_length,
-                                int *clause_weights, int *class_sum, int *X) {
+                                float *clause_weights, float *class_sum, int *X) {
     int index = blockIdx.x * blockDim.x + threadIdx.x;
     int stride = blockDim.x * gridDim.x;
 
@@ -91,7 +91,7 @@ __global__ void evaluate_packed(unsigned int *included_literals, unsigned int *i
 
         if (clause_output) {
             for (int class_id = 0; class_id < CLASSES; ++class_id) {
-                int clause_weight = clause_weights[class_id * CLAUSES + clause];
+                float clause_weight = clause_weights[class_id * CLAUSES + clause];
                 atomicAdd(&class_sum[class_id], clause_weight);
             }
         }
